Updates to code
It's been a while since I have worked on this project, leaguepedia parser has been updated to version 2.0 so I decided to redo my code to make it hopefully cleaner and easier to read. I've removed Team data as I found it was not working as I intended instead only general champion data is now here.